requirements.txt
openai>=1.0.0
anthropic==0.8.1
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
Pillow==10.2.0
PyGithub==2.1.1
vercel==0.2.1
speechrecognition==3.10.1
pyaudio==0.2.14
soundfile==0.12.1
sounddevice==0.4.6
scrapybara==0.1.0
google-ai-generative-language==0.2.0
google-generativeai==0.3.2
elevenlabs==0.3.0
